fix: change weightChunk type from int8_t to int32_t address issue Potential Type Error #9

+++ BitNetMCU_inference.c
@@ -153,7 +153,7 @@ void processfclayer( int8_t *activations,  const uint32_t *weights, int32_t bits
             }
         } else if (bits_per_weight == 8 + 8 ) {   // 8 bit twos-complement
             for (uint32_t k = 0; k < n_input; k+=4) {
-                int8_t weightChunk = *weightidx++;
+                int32_t weightChunk = *weightidx++;
                 for (uint32_t j = 0; j < 4; j++) {
                     int32_t in=*activations_idx++;
                     int32_t weight = (weightChunk) >> (32-8); // extend sign, cut off lower bits
